bot/requirements.txt
beautifulsoup4==4.12.3
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
selenium==4.21.0
tldextract==5.1.1
whois==0.9.24
joblib==1.4.2
python-whois
dnspython
